Rename harness::call to harness::trigger.
Aligns the browser ingestion bridge with iii.trigger naming; updates console/web to invoke the new function id.
