+++ harness-node/src/harness/register.ts
@@ -1,6 +1,6 @@
 import { loadConfig } from '../runtime/config.js';
 import type { ISdk } from '../runtime/iii.js';
-import { register as registerCall } from './call.js';
+import { register as registerTrigger } from './trigger.js';
 import { loadHarnessConfig } from './config.js';
 import { spawnPumps } from './fanout/index.js';
 import { register as registerFs } from './fs.js';
@@ -11,7 +11,7 @@ import { FanoutState, registerSubscriptions } from './ui-subscribe.js';
 export async function register(iii: ISdk, ctx: { configPath: string; url: string }): Promise<void> {
   const cfg = await loadConfig(ctx.configPath);
   const harness = loadHarnessConfig(cfg);
-  registerCall(iii);
+  registerTrigger(iii);
   const fanoutState = new FanoutState();
   registerSubscriptions(iii, fanoutState);
   spawnPumps(iii, fanoutState);
